requirements.txt
argparse
Babel>=1.3
iso8601>=0.1.9
netaddr>=0.7.6
oslo.config>=1.2.0
pbr>=0.6,!=0.7,<1.0
PrettyTable>=0.7,<0.8
requests>=1.1
six>=1.6.0
stevedore>=0.14
